- Update Mongoose Middleware

+++ src/domains/genre/model/middleware/GenreQueryMiddleware.ts
@@ -5,5 +5,7 @@ import Movie from "../../../movie/model/Movie.js";
 
 export async function DeleteGenreQueryPreMiddleware(this: Query<any, IGenre>) {
     const {_id} = this.getFilter();
+    if (!_id) return;
+
     await Movie.deleteMany({genre: _id}).exec();
 }
